+++ src/config.php
@@ -24,8 +24,10 @@ return [
     // Phone number to use for SMS testing purposes
     //'testToPhoneNumber' => getenv('TEST_TO_PHONE_NUMBER'),
 
-    // Adjust the default Twig sandbox configuration
+    // Customize the Twig sandbox configuration
     // https://plugins.doublesecretagency.com/notifier/messages/twig-sandbox
-    //'twigSandbox' => []
+    //'twigSandboxMode' => 'append',
+    //'twigSandboxBlacklist' => [],  // Do not use both lists
+    //'twigSandboxWhitelist' => []   // at the same time
 
 ];